refactor(directory_structure): rename and relocate RAG server
Renamed `rag_pipeline.py` to `dir_rag_server.py` for improved clarity and consistency with the module's purpose. Updated import paths across affected files and relocated related tests accordingly. Simplified the project structure to enhance maintainability.
