+++ src/mcp_llamaindex/mcp_server.py
@@ -1,6 +1,6 @@
 import logging
 
-from mcp_llamaindex.rag_pipeline import DirectoryRagServer
+from mcp_llamaindex.dir_rag_server import DirectoryRagServer
 
 # Initialize FastMCP server
 mcp = DirectoryRagServer().as_server()
